LayoutTests/fast/forms/image/width-and-height-of-detached-input.html

<!DOCTYPE html>
<script src="../../../resources/js-test-pre.js"></script>
<script>
description('Width and height properties of a detached input element should return 0.');
var input = document.createElement('input');
input.type = 'image';
shouldBe('input.width', '0');
shouldBe('input.height', '0');
</script>
<script src="../../../resources/js-test-post.js"></script>
