third_party/blink/web_tests/fast/forms/image/width-and-height-of-detached-input-expected.txt

Width and height properties of a detached input element should return 0.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.width is 0
PASS input.height is 0
PASS successfullyParsed is true

TEST COMPLETE
